feat: form fields are now reset upon submission
