+++ packages/fe/pages/apply/large.vue
@@ -302,6 +302,7 @@ export default {
         const application = await this.$form.applyFormToSchema(formId, this.application)
         this.$form.clearLocalStorage(formId)
         await this.submitApplication({ application, bytes, thresholds })
+        this.$form.reset(formId)
       }
       this.$button('application-submit-button').set({ loading: false })
     }
